Modelo predictedRoadTypes renombrado. Funcionalidad compartida relacionada a Tracks movida a controller de Tracks.
