+++ src/controllers/sumarizations.ts
@@ -6,25 +6,16 @@ import {
     map
 } from 'rxjs/operators';
 import {
-    getTracksMapByCity
+    getTracksMapByCity, sumarizeTracksByCity
 } from './tracks';
 import {
     ISumarizingObject,
-    ISumarizationSegment,
     ISumarizedObject,
-    mapRangeToSumarizingSegment,
-    matches
 } from '../interfaces/Segment';
 import {
     IRange,
     ITrack
 } from '../interfaces/Track';
-
-// TODO: Refactorizar esto... deberia ir en un archivo aparte de configuracion o de constantes
-// TODO: Agregar funcion que calcule peso de forma dinamica haciendo una resta entre
-// TODO: Date() y el date del segmento
-const NEW_DATA_WEIGHT = 0.6;
-const OLD_DATA_WEIGHT = 1 - NEW_DATA_WEIGHT;
 
 const errorCallback = (err: any) => {
     console.error(err);
@@ -65,57 +56,6 @@ export const sumarizeTracksCallback = (req: any, res: any): void => {
         .subscribe((sumarizations: ISumarizedObject[]) => {
             putSumarizationsCallback(req, res, sumarizations);
         }, errorCallback);
-}
-
-const sumarizeTracksByCity = (items: ISumarizingObject[]): ISumarizedObject[] => {
-    return items.map((item: ISumarizingObject) => sumarizeTracks(item));
-}
-
-//? 3) Ejecuto la sumarizacion de Tracks de una Ciudad
-const sumarizeTracks = (item: ISumarizingObject): ISumarizedObject => {
-    const date = Date.parse(new Date().toDateString());
-    const sumarizedSegments: ISumarizationSegment[] = [];
-
-    item.tracks.forEach((track: ITrack) => {
-        sumarizeNextTrack(sumarizedSegments, track);
-    });
-
-    return <ISumarizedObject>{
-        cityId: item.cityId,
-        date,
-        ranges: sumarizedSegments
-    };
-}
-
-const sumarizeNextTrack = (array: ISumarizationSegment[], track: ITrack): void => {
-    let ranges: IRange[] = track.ranges;
-    let segments: ISumarizationSegment[] = [];
-    
-    segments = ranges.map((r: IRange) => mapRangeToSumarizingSegment(r));
-
-    segments.forEach((s: ISumarizationSegment) => {
-        addSegment(s, array);
-    });
-}
-
-const addSegment = (sNew: ISumarizationSegment, array: ISumarizationSegment[]): void => {
-    const matchingSegment = findMatchingSegment(sNew, array);
-    if (matchingSegment) {
-        updateMatchingSegment(sNew, matchingSegment);
-    } else {
-        sNew.accuracy = 1;
-        array.push(sNew);
-    }
-}
-
-const findMatchingSegment = (mySegment: ISumarizationSegment, array: ISumarizationSegment[]): ISumarizationSegment | undefined => {
-    return array.find((s: ISumarizationSegment) => matches(mySegment, s));
-}
-
-const updateMatchingSegment = (segment: ISumarizationSegment, matchingSegment: ISumarizationSegment): void => {
-    matchingSegment.score = matchingSegment.score * OLD_DATA_WEIGHT + segment.score * NEW_DATA_WEIGHT;
-    matchingSegment.date = segment.date;
-    matchingSegment.accuracy++;
 }
 
 const getSumarizationsByFilter = async (filter: {}) => {
